requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
streamlit==1.48.1
python-dotenv==1.0.1
requests==2.32.4
plotly==5.18.0
pandas==2.1.4
google-generativeai==0.3.2
transformers==4.36.2
torch==2.1.2